added generic client class, moved retry function to class and made ingest and upload clients extend new class
