+++ src/utils/upload-client/upload-client.ts
@@ -1,35 +1,32 @@
 import Promise from "bluebird";
 import request from "request-promise";
 import {FileValidationRequest, UploadApiConnectionProperties} from "../../common/types";
+import {Client} from "../client";
 
-class UploadClient {
-    uploadApiConnectionProperties: UploadApiConnectionProperties;
-    uploadUrl: string;
-    uploadApiKey: string;
+class UploadClient extends Client {
 
     constructor(uploadApiConnectionProperties: UploadApiConnectionProperties, uploadApiKey: string) {
-        this.uploadApiConnectionProperties = uploadApiConnectionProperties;
-        this.uploadUrl = uploadApiConnectionProperties["scheme"] + "://" + uploadApiConnectionProperties["host"] + ":" + uploadApiConnectionProperties["port"];
-        this.uploadApiKey = uploadApiKey;
+        const baseUrl =  uploadApiConnectionProperties["scheme"] + "://" + uploadApiConnectionProperties["host"] + ":" + uploadApiConnectionProperties["port"];
+        super(baseUrl, uploadApiKey);
     }
 
     requestFileValidationJob(validationRequest: FileValidationRequest) : Promise<string> {
-        return UploadClient._requestFileValidationJob(validationRequest, this.uploadUrl, this.uploadApiKey);
+        return this.retry(5, this._requestFileValidationJob.bind(this), [validationRequest], "Retrying requestFileValidationJob()");
     }
 
-    static _requestFileValidationJob(validationRequest: FileValidationRequest, uploadUrl: string, uploadApiKey: string) : Promise<string> {
+    _requestFileValidationJob(validationRequest: FileValidationRequest) : Promise<string> {
         return new Promise<string>((resolve, reject) => {
             const uploadAreaUuid = validationRequest.uploadAreaUuid;
             const fileName = validationRequest.fileName;
             const imageUrl = validationRequest.validationImageUrl;
 
-            const validateFileUrl = uploadUrl + "/v1/area/" + uploadAreaUuid + "/" + encodeURIComponent(fileName) + "/validate";
+            const validateFileUrl = this.clientBaseUrl + "/v1/area/" + uploadAreaUuid + "/" + encodeURIComponent(fileName) + "/validate";
             request({
                 method: "PUT",
                 url: validateFileUrl,
                 json: true,
                 headers: {
-                    "Api-key" : uploadApiKey
+                    "Api-key" : this.clientApiKey
                 },
                 body: {
                     "validator_image": imageUrl
